backend/requirements.txt
openai
google-generativeai
trafilatura
python-dotenv
pydantic
requests
httpx
fastapi
uvicorn
tenacity
